+++ ext/cdict/table.h
@@ -27,7 +27,7 @@
 typedef struct T *T;
 extern T table_new(int hint,
   int cmp(const void *x, const void *y),
-  unsigned hash(const void *key));
+  unsigned long hash(const void *key));
 extern void table_free(T *table);
 extern int table_length(T table);
 extern void *table_put(T table, const void *key,
